requirements.txt
pytest-allure-adaptor==1.7.7
humanize==0.5.1
pyhocon==0.3.38
pytest==3.1.0
pytest-allure-adaptor==1.7.7
requests==2.18.1
requests-mock==1.3.0
sure==1.4.5